+++ src/extension.ts
@@ -6,6 +6,7 @@ import * as vscode from 'vscode';
 // import { error } from 'console';
 import { arquiveFolder, versionFolder, zipArquive } from './folderComands';
 import { createStructure } from './folderStruct';
+import { nodePrj, simpleDesktopPrj, desktopPrj } from './folderProjects';
 
 
 // This method is called when your extension is activated
@@ -33,12 +34,27 @@ export function activate(context: vscode.ExtensionContext) {
     let disp4 = vscode.commands.registerCommand('extension.zipArquive', async (uri: vscode.Uri) => {
 		zipArquive(context, uri, outputChannel);
 	});
+
+    let disp5 = vscode.commands.registerCommand('extension.nodePrj', async (uri: vscode.Uri) => {
+		nodePrj(context, uri, outputChannel);
+	});
+
+    let disp6 = vscode.commands.registerCommand('extension.simpleDeskPrj', async (uri: vscode.Uri) => {
+		simpleDesktopPrj(context, uri, outputChannel);
+	});
+
+    let disp7 = vscode.commands.registerCommand('extension.deskPrj', async (uri: vscode.Uri) => {
+		desktopPrj(context, uri, outputChannel);
+	});
   
 	context.subscriptions.push(outputChannel);
 	context.subscriptions.push(disp1);
 	context.subscriptions.push(disp2);
 	context.subscriptions.push(disp3);
 	context.subscriptions.push(disp4);
+	context.subscriptions.push(disp5);
+	context.subscriptions.push(disp6);
+	context.subscriptions.push(disp7);
 }
 
 // This method is called when your extension is deactivated